requirements.txt
fastapi
uvicorn[standard]
python-mistralai
python-dotenv
aiofiles
python-docx
pydantic
requests
